NationAI: Finish sea based action by issuing final order.
This reenables code that appears to have been fully finished. I believe
it was disabled due to an issue late in development that caused the
final order to fail. I fixed an issue with unit move blocked by team in
the previous commit. Now when this sea code is enabled, it appears to be
working as expected.

Previously, when the AI landed the people, the action was not
necessarily ended. Eventually it could be reprocessed or reordered, but
it took some time. This fix makes the AI act immediately, which is what
we want.