Vga: Fix mouse bound checks.
When scrolling to the edge with the improper checking, this was causing a
loop of warping and mouse move events. With improvements in SDL, it
increased the load on the CPU and caused more warps per frame, making the
mouse pointer to stick to the edge longer going throught the processing.