Town: Initialize statistics for new nation.
If population is left zero, then there is a small window until the next
month that the nation will immediately surrender if Nation::think_surrender()
is executed, because this is used as a defeat condition.

7k2 tried to fix this with a couple changes. It tried to switch init
order in a risky (buggy?) way, and created a work-around as well, (why?).
The probability of Nation::think_surrender() of happening under recent
event conditions should be a clue something else was amiss.